requirements.txt
streamlit
openai
Faker
fastapi
requests
uvicorn
tiktoken
python-dotenv
